Add files via upload
Light and Dark Mode Switch GitHub repository! This project showcases a sleek and modern implementation of a light and dark mode toggle using HTML, CSS, and JavaScript. With this repository, you can learn how to create an engaging user experience by seamlessly transitioning between two captivating themes.

The project includes a visually appealing card designed with inspiration from MasterCard. Additionally, there is a small button adorned with a Font Awesome icon, adding a touch of style and elegance to the interface.

By default, the project is set to a refreshing light theme, providing a clean and vibrant user interface. However, with a simple click of the button, the theme transforms into an enchanting dark mode. The background, color scheme, and icon dynamically adjust to create a visually pleasing and cohesive aesthetic.

Whether you're a beginner or an experienced developer, this repository offers valuable insights into building a light and dark mode switch for your own web applications.

Note: Feel free to customize the project further or incorporate it into your existing projects to provide a seamless theme switching functionality.